connectors

connectors

draws bezier connectors between pairs of signed points points a = (x0, y0, s0) and b(x1, y1, s1)

"S" connectors will use bezier control points that are located in an intermediate y location between y0 and y1, and "U" connectors will use bezier control points that are located in an intermediate x location between x0 and x1.

type can be "U" or "S", if "S", then sign of corresponding v value is ignored (ie control points will always be vertically in between a and b, in addition, neither intermeidate control will never be more than 1/3 of the y distance between a and b

The sign of the endpoints refers to the direction from which the connector approaches the point, s0 = 1 (or '+') means from the right and s0 = 0 means from the left, similarly for s1

h = horizonal "bulge" of signed connector, ie depending on the sign the extent to which the connector will extend to the left or the right of points an and b v = vertical bulge (only applies to U connectors), how far above max(y0, y1) the U connector will bulge if v>0 or how far below min(y0, y1) the connector will bulge.

nsteps determines how many segments used to approximate the curve

all args except nsteps can be vecorized

connectors(
  x0,
  y0,
  s0 = 1,
  x1,
  y1,
  s1 = 1,
  v = 0.1,
  h = 0.1,
  type = "S",
  f.arrow = T,
  b.arrow = F,
  nsteps = 100,
  cex.arrow = 1,
  col.arrow = "black",
  lwd = 1,
  lty = 1,
  col = "black"
)

Author

Marcin Imielinski