draw.grl

wrapper function around draw.ranges to draw GRangesList across a set of genomic windows By default, these windows correspond to the smallest set of "covered" regions (ie containing at least one grl in their span, but can also be specified as an argument.

basically same idea as draw.granges, except can more compactly draw / stack paired / group reads or gene models / alignments and deal with "grouped" intervals

additional values features of input grl that can impact drawing: ywid = vertical thickness of y segments border.col = color border (can override colormap)

if optional arg draw.var == T will either call varbase on grl (e.g. if grl is a GRanges or GappedAlignment object) to get variant ranges or will take vars object (GRangesList that is of same length as grl) and will display variant ranges according to color scheme in arg "var.col" "var.col" is a named vector that maps variant types to colors, with the following syntax:

XA, XT, XG, XC --> colors of base substitutions S, D, I --> colors of sooutft clipped, deletions, insertions e.g. var.col = c(XA = 'blue', S = 'yellow', 'I' = green') would replace the default variant coloring of "A" substitutions, soft clipped regions, and insertions with blue, yellow, and green colors, respectively.

draw.grl(
  grl,
  y = NULL,
  ylim = NULL,
  ylim.subplot = NULL,
  ywid = NULL,
  edges = NULL,
  draw.paths = F,
  draw.var = F,
  var = NULL,
  var.col = NULL,
  var.soft = T,
  windows,
  win.gap = NULL,
  stack.gap,
  min.gapwidth = 1,
  col = NULL,
  border = NA,
  col.backbone = alpha("gray", 0.8),
  gr.colorfield = NULL,
  gr.colormap = NULL,
  gr.labelfield = NULL,
  grl.labelfield = NULL,
  grl.colorfield = NULL,
  grl.cexfield = NULL,
  leg.params,
  labels = NULL,
  labels.suppress = F,
  labels.suppress.grl = labels.suppress,
  labels.suppress.gr = labels.suppress,
  spc.label = 0.05,
  adj.label = c(0, 1),
  cex.label = 1,
  gr.cex.label = 0.8 * cex.label,
  gr.srt.label = 0,
  gr.adj.label = c(0, 0.5),
  new.plot,
  new.axis,
  sep.lty = 2,
  sep.lwd = 1,
  sep.bg.col = "gray95",
  sep.draw = TRUE,
  y.pad,
  xaxis.prefix = "",
  xaxis.suffix = "MB",
  xaxis.unit = 1,
  xaxis.round = 3,
  xaxis.interval = "auto",
  xaxis.pos = 1,
  xaxis.pos.label,
  xaxis.cex.label,
  xaxis.newline = FALSE,
  xaxis.chronly = FALSE,
  xaxis.ticklen = 1,
  xaxis.width = TRUE,
  xaxis.label.angle = 0,
  xaxis.cex.tick = 1,
  ylim.grid = ylim,
  y.grid = NA,
  ylab = NULL,
  y.grid.col = alpha("gray", 0.5),
  y.grid.pretty = 5,
  y.grid.cex = 1,
  y.grid.lty = 2,
  y.grid.lwd = 1,
  path.col = "black",
  path.lwd = 1,
  path.col.arrow = path.col,
  path.cex.arrow = 1,
  path.stack.y.gap = 1,
  path.stack.x.gap = 0,
  path.cex.v = 1,
  path.cex.h = 1,
  draw.backbone = NULL,
  xlim = c(0, 20000),
  points = NA,
  circles = FALSE,
  bars = FALSE,
  y0.bar = NULL,
  lines = F,
  angle,
  verbose = FALSE,
  triangle = FALSE,
  ylim.parent = NULL,
  legend.params = list(plot = TRUE),
  bg.col = NA,
  ...
)

Author

Marcin IMielinski